refactor: move RecoveryPhraseModal
