+++ front-end/src/renderer/components/modals/RecoveryPhraseModal.vue
